Isolate hello-compute examples to own file
